ex5 fix

+++ src/exercise5.c
@@ -5,18 +5,18 @@ int main(int argc, char** argv) {
     const int n = 3;
     const int q = 10;
 
-    int m1[m][n];
-    int m2[n][q];
-    int m3[m][q];
+    double m1[m][n];
+    double m2[n][q];
+    double m3[m][q];
  
     for (int i = 0; i < m; i++) {
         for (int j = 0; j < n; j++) {
-            scanf("%d", &m1[i][j]);
+            scanf("%lf", &m1[i][j]);
         }
     }
     for (int i = 0; i < n; i++) {
         for (int j = 0; j < q; j++) {
-            scanf("%d", &m2[i][j]);
+            scanf("%lf", &m2[i][j]);
         }
     }
 
@@ -30,7 +30,7 @@ int main(int argc, char** argv) {
     }
     for (int i = 0; i < m; i++) {
         for (int j = 0; j < q; j++) {
-            printf("%d ", m3[i][j]);
+            printf("%.2lf ", m3[i][j]);
         }
     }
     return 0;
